change base sth & module sth
device
    修改命名规范
    修改设备的结构体定义
    增加构造函数，并在构造时进行判空处理
    优化头文件结构
timer
    在头文件中引出定时器相关宏定义
    将延时参数统一修改为uint64_t
    删除宏函数，统一使用内联
protocol
    修改命名
oled&ds18b20
    随device修改初始化结构
    将oled的字库修改为utf8-BOM编码，此编码下使用汉字不会出现warning
    因为编码修改，将oled的printchinese函数及中文数组函数的长度由2增加至3

+++ timer.c
@@ -38,7 +38,7 @@ float TIMER_fps(void) {
 //获取当前时间, 记录当前时间, 更新state
 //state1.输入:时间 返回:0/1(并复位state)
 //获取当前时间, 比较历史时间
-int8_t TIMER_mscmptor(uint16_t ms, volatile uint64_t *compare, volatile int8_t *state) {
+int8_t TIMER_mscmptor(uint64_t ms, volatile uint64_t *compare, volatile int8_t *state) {
     extern volatile uint64_t time_us;
     extern volatile int8_t flag_timerrupt;
     uint64_t time = 0;
@@ -62,7 +62,7 @@ int8_t TIMER_mscmptor(uint16_t ms, volatile uint64_t *compare, volatile int8_t *
     }
     return result;
 }
-int8_t TIMER_scmptor(uint16_t s, volatile uint64_t *compare, volatile int8_t *state) {
+int8_t TIMER_scmptor(uint64_t s, volatile uint64_t *compare, volatile int8_t *state) {
     extern volatile uint64_t time_us;
     extern volatile int8_t flag_timerrupt;
     uint64_t time = 0;
